+++ ts/apps/polaris-cli/src/app/tasks/promise-task.ts
@@ -1,0 +1,14 @@
+import { Task } from './task';
+
+/**
+ * Executes a promise.
+ */
+ export class PromiseTask implements Task {
+
+    constructor(public fn: () => Promise<void>) { }
+
+    execute(): Promise<void> {
+        return this.fn();
+    }
+
+}