+++ ts/libs/polaris-nx/src/generators/config/schema.d.ts
@@ -1,0 +1,6 @@
+export interface ConfigurationSchema {
+    modifier: string;
+    project: string;
+    key: string;
+    value?: string;
+}